refactor: move emitOperation function
