+++ packages/serverless-typespec-generator/src/ir/emit.ts
@@ -1,5 +1,13 @@
-import { emitOperation, renderType } from "../typespec/ir/emit"
-import type { AliasIR, ModelIR, TypeSpecIR } from "../typespec/ir/type"
+import { renderType } from "../typespec/ir/emit"
+import {
+  type AliasIR,
+  type HttpResponseIR,
+  type ModelIR,
+  type OperationIR,
+  type TypeSpecIR,
+  isHttpResponse,
+  isHttpResponses,
+} from "../typespec/ir/type"
 
 export function emitTypeSpec(irList: TypeSpecIR[]): string {
   const lines: string[] = []
@@ -51,4 +59,42 @@ export function emitModel(model: ModelIR): string {
   lines.push("}")
 
   return lines.join("\n")
+}
+
+export function emitOperation(operation: OperationIR): string {
+  const lines: string[] = []
+
+  const parameters: string[] = []
+  const paramEntries = Object.entries(operation.parameters ?? {})
+  const pathParams = new Set(operation.http?.params ?? [])
+  for (const [name, prop] of paramEntries) {
+    const decorator = pathParams.has(name) ? "@path" : ""
+    const optional = prop.required ? "" : "?"
+    const type = renderType(prop.type)
+    parameters.push(`${decorator} ${name}${optional}: ${type}`)
+  }
+
+  if (operation.requestBody) {
+    parameters.push(`@body body: ${renderType(operation.requestBody)}`)
+  }
+
+  let returnType = "void"
+  if (isHttpResponses(operation.returnType)) {
+    returnType = operation.returnType.map(renderHttpResponse).join(" | ")
+  } else if (isHttpResponse(operation.returnType)) {
+    returnType = renderHttpResponse(operation.returnType)
+  } else if (operation.returnType) {
+    returnType = renderType(operation.returnType)
+  }
+
+  lines.push(`@route("${operation.route}")`)
+  lines.push(`@${operation.method}`)
+  lines.push(`op ${operation.name}(${parameters.join(", ")}): ${returnType};`)
+
+  return lines.join("\n")
+}
+
+function renderHttpResponse(r: HttpResponseIR): string {
+  const body = renderType(r.body)
+  return `{ @statusCode statusCode: ${r.statusCode}; @body body: ${body} }`
 }
