+++ packages/serverless-typespec-generator/src/typespec/ir/emit.test.ts
@@ -1,9 +1,4 @@
-import { describe, expect, it } from "vitest"
-
 import { formatTypeSpec } from "@typespec/compiler"
-import dedent from "dedent"
-import { emitOperation } from "./emit"
-import type { OperationIR } from "./type"
 
 async function normalizeTypeSpec(code: string) {
   const formattedCode = await formatTypeSpec(code, {
@@ -13,218 +8,3 @@ async function normalizeTypeSpec(code: string) {
   })
   return formattedCode.trimEnd()
 }
-
-describe("emitOperation", () => {
-  it("should emit a simple operation", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "createUser",
-      method: "post",
-      route: "/users",
-      requestBody: {
-        name: { type: "string", required: true },
-        email: { type: "string", required: true },
-      },
-      returnType: {
-        id: { type: "string", required: true },
-        name: { type: "string", required: true },
-        email: { type: "string", required: true },
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users")
-      @post
-      op createUser(
-        @body body: {
-          name: string;
-          email: string;
-        },
-      ): {
-        id: string;
-        name: string;
-        email: string;
-      };
-    `)
-  })
-  it("should emit an operation with no request body", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUser",
-      method: "get",
-      route: "/users/{id}",
-      returnType: {
-        id: { type: "string", required: true },
-        name: { type: "string", required: true },
-        email: { type: "string", required: true },
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users/{id}")
-      @get
-      op getUser(): {
-        id: string;
-        name: string;
-        email: string;
-      };
-    `)
-  })
-  it("should emit an operation with array response", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUsers",
-      method: "get",
-      route: "/users",
-      returnType: [
-        {
-          id: { type: "string", required: true },
-          name: { type: "string", required: true },
-          email: { type: "string", required: true },
-        },
-      ],
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users")
-      @get
-      op getUsers(): {
-        id: string;
-        name: string;
-        email: string;
-      }[];
-    `)
-  })
-  it("should emit an operation with external model", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUser",
-      method: "get",
-      route: "/users/{id}",
-      returnType: {
-        ref: "User",
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users/{id}")
-      @get
-      op getUser(): User;
-    `)
-  })
-  it("should emit an operation with union model", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUser",
-      method: "get",
-      route: "/users/{id}",
-      returnType: {
-        union: [
-          {
-            id: { type: "string", required: true },
-            name: { type: "string", required: true },
-            email: { type: "string", required: true },
-          },
-          {
-            code: { type: "string", required: true },
-            message: { type: "string", required: true },
-          },
-        ],
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users/{id}")
-      @get
-      op getUser(): {
-        id: string;
-        name: string;
-        email: string;
-      } | {
-        code: string;
-        message: string;
-      };
-    `)
-  })
-  it("should emit an operation with http response", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUser",
-      method: "get",
-      route: "/users/{id}",
-      returnType: {
-        statusCode: 200,
-        body: {
-          id: { type: "string", required: true },
-          name: { type: "string", required: true },
-          email: { type: "string", required: true },
-        },
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users/{id}")
-      @get
-      op getUser(): {
-        @statusCode statusCode: 200;
-        @body body: {
-          id: string;
-          name: string;
-          email: string;
-        };
-      };
-    `)
-  })
-  it("should emit an operation with http response with array", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUsers",
-      method: "get",
-      route: "/users",
-      returnType: {
-        statusCode: 200,
-        body: [
-          {
-            id: { type: "string", required: true },
-            name: { type: "string", required: true },
-            email: { type: "string", required: true },
-          },
-        ],
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users")
-      @get
-      op getUsers(): {
-        @statusCode statusCode: 200;
-        @body body: {
-          id: string;
-          name: string;
-          email: string;
-        }[];
-      };
-    `)
-  })
-  it("should emit an operation with path parameters", async () => {
-    const operation: OperationIR = {
-      kind: "operation",
-      name: "getUser",
-      method: "get",
-      route: "/users/{id}",
-      parameters: {
-        id: { type: "string", required: true },
-      },
-      returnType: { ref: "User" },
-      http: {
-        params: ["id"],
-      },
-    }
-    const result = emitOperation(operation)
-    expect(await normalizeTypeSpec(result)).toBe(dedent`
-      @route("/users/{id}")
-      @get
-      op getUser(@path id: string): User;
-    `)
-  })
-})
